refactored stdlib so that modules are in separate files
