+++ include/pico/stdlib/extra.h
@@ -1,9 +1,7 @@
-#ifndef __PICO_VALUES_STDLIB_H
-#define __PICO_VALUES_STDLIB_H
+#ifndef __PICO_STDLIB_EXTRA_H
+#define __PICO_STDLIB_EXTRA_H
 
-#include "platform/jump.h"
 #include "data/stream.h"
-
 #include "pico/values/modular.h"
 
 // Hooks
@@ -11,6 +9,7 @@ void set_exit_callback(jump_buf* buf);
 
 // Set the default value of dynamic variables
 
+Allocator* get_std_allocator();
 Module* get_std_current_module();
 Module* set_std_current_module(Module* al);
 
@@ -18,15 +17,9 @@ void set_current_package(Package* current);
 void set_std_istream(IStream* current);
 void set_std_ostream(OStream* current);
 
-PiType* get_syntax_type();
-PiType* get_array_type();
-PiType* get_c_type();
-
 Allocator* get_std_tmp_allocator();
 Allocator* set_std_tmp_allocator(Allocator* al);
 
-Allocator* get_std_allocator();
-
-Package* base_package(Assembler* ass, Allocator* default_allocator, Allocator* a);
+void add_extra_module(Assembler* ass, Package* base, Allocator* default_allocator, Allocator* a);
 
 #endif
