+++ include/pico/stdlib/num.h
@@ -1,0 +1,8 @@
+#ifndef __PICO_STDLIB_NUM_H
+#define __PICO_STDLIB_NUM_H
+
+#include "pico/values/modular.h"
+
+void add_num_module(Assembler* ass, Package* base, Allocator* a);
+
+#endif
