+++ include/pico/stdlib/stdlib.h
@@ -1,0 +1,8 @@
+#ifndef __PICO_STDLIB_STDLIB_H
+#define __PICO_STDLIB_STDLIB_H
+
+#include "pico/values/modular.h"
+
+Package* base_package(Assembler* ass, Allocator* default_allocator, Allocator* a);
+
+#endif
