+++ include/pico/stdlib/user.h
@@ -1,0 +1,8 @@
+#ifndef __PICO_STDLIB_USER_H
+#define __PICO_STDLIB_USER_H
+
+#include "pico/values/modular.h"
+
+void add_user_module(Package* base, Allocator* a);
+
+#endif
