+++ src/pico/eval/call.c
@@ -5,7 +5,7 @@
 
 #include "pico/eval/call.h"
 #include "pico/values/types.h"
-#include "pico/values/stdlib.h"
+#include "pico/stdlib/extra.h"
 #include "pico/codegen/codegen.h"
 
 EvalResult pico_run_toplevel(TopLevel top, Target target, LinkData links, Module* module, Allocator* a, ErrorPoint* point) {
